feat(cli): wire harness maintenance run subcommand
