+++ packages/cli/src/commands/maintenance.ts
@@ -13,101 +13,16 @@
 
 import { Command } from 'commander';
 import * as path from 'node:path';
-import * as fs from 'node:fs';
 import chalk from 'chalk';
-import {
-  BUILT_IN_TASKS,
-  TaskOutputStore,
-  WorkflowLoader,
-  type TaskDefinition,
-} from '@harness-engineering/orchestrator';
-import type { CustomTaskDefinition, MaintenanceConfig } from '@harness-engineering/types';
+import { TaskOutputStore } from '@harness-engineering/orchestrator';
 import type { PersistedOutputEntry } from '@harness-engineering/orchestrator';
 import { ExitCode } from '../utils/errors';
-import { logger } from '../output/logger';
+import { loadMaintenanceConfig, mergeResolvedTasks } from './maintenance-config';
+import { runMaintenanceRun } from './maintenance-run';
 
-export async function loadMaintenanceConfig(cwd: string): Promise<MaintenanceConfig | null> {
-  const workflowPath = path.join(cwd, 'harness.orchestrator.md');
-  if (!fs.existsSync(workflowPath)) return null;
-  const loader = new WorkflowLoader();
-  const result = await loader.loadWorkflow(workflowPath);
-  if (!result.ok) return null;
-  // Spec B Phase 2 / S3: surface non-blocking routing warnings at startup.
-  for (const w of result.value.warnings) logger.warn(w);
-  return (result.value.config as { maintenance?: MaintenanceConfig }).maintenance ?? null;
-}
-
-export function mergeResolvedTasks(config: MaintenanceConfig | null): TaskDefinition[] {
-  const overrides = config?.tasks ?? {};
-  const tasks: TaskDefinition[] = [];
-  appendBuiltIns(tasks, overrides);
-  appendCustomTasks(tasks, config?.customTasks ?? {}, overrides);
-  return tasks;
-}
-
-function appendBuiltIns(
-  tasks: TaskDefinition[],
-  overrides: NonNullable<MaintenanceConfig['tasks']>
-): void {
-  for (const t of BUILT_IN_TASKS) {
-    const ov = overrides[t.id];
-    if (ov?.enabled === false) continue;
-    const next: TaskDefinition = { ...t };
-    if (ov?.schedule !== undefined) next.schedule = ov.schedule;
-    tasks.push(next);
-  }
-}
-
-function appendCustomTasks(
-  tasks: TaskDefinition[],
-  customs: Record<string, CustomTaskDefinition>,
-  overrides: NonNullable<MaintenanceConfig['tasks']>
-): void {
-  for (const [id, def] of Object.entries(customs)) {
-    const ov = overrides[id];
-    if (ov?.enabled === false) continue;
-    tasks.push(buildCustomTaskDefinition(id, def, ov?.schedule));
-  }
-}
-
-function buildCustomTaskDefinition(
-  id: string,
-  def: CustomTaskDefinition,
-  scheduleOverride: string | undefined
-): TaskDefinition {
-  const out: TaskDefinition = {
-    id,
-    type: def.type,
-    description: def.description,
-    schedule: scheduleOverride ?? def.schedule,
-    branch: def.branch,
-    isCustom: true,
-  };
-  copyOptional(def, out as unknown as Record<string, unknown>, [
-    'checkCommand',
-    'checkScript',
-    'fixSkill',
-    'inlineSkills',
-    'inlineSkillsBudgetTokens',
-    'contextFrom',
-    'contextFromMaxAgeMinutes',
-    'outputRetention',
-    'costCeiling',
-  ]);
-  return out;
-}
-
-function copyOptional(
-  src: Record<string, unknown> | object,
-  dst: Record<string, unknown>,
-  keys: string[]
-): void {
-  const s = src as Record<string, unknown>;
-  for (const k of keys) {
-    const v = s[k];
-    if (v !== undefined) dst[k] = v;
-  }
-}
+// Re-export the task-resolution helpers (promoted to named exports for the
+// `run` subcommand) so existing importers of `./maintenance` keep working.
+export { loadMaintenanceConfig, mergeResolvedTasks } from './maintenance-config';
 
 interface ListRow {
   id: string;
@@ -175,6 +90,30 @@ export function createMaintenanceCommand(): Command {
     .action(async (taskId, opts) => {
       await runShow(taskId, opts);
       process.exit(ExitCode.SUCCESS);
+    });
+
+  command
+    .command('run [taskId...]')
+    .description('Run overdue (default) / selected maintenance tasks; report-first unless --fix')
+    .option('--all', 'Run all sweep-eligible tasks (not just overdue)', false)
+    .option('--only <ids>', 'Comma-separated task ids to run')
+    .option('--skip <ids>', 'Comma-separated task ids to exclude')
+    .option('--fix', 'Dispatch fixes per task type (default: report-only)', false)
+    .option('--concurrency <n>', 'Max parallel tasks (report mode); --fix forces 1')
+    .option('--json', 'Emit machine-readable consolidated report', false)
+    .option('--path <path>', 'Project root path', '.')
+    .action(async (taskIds: string[], opts) => {
+      const cwd = path.resolve(opts.path);
+      const result = await runMaintenanceRun(cwd, {
+        positional: taskIds,
+        all: opts.all,
+        only: opts.only,
+        skip: opts.skip,
+        fix: opts.fix,
+        concurrency: opts.concurrency,
+        json: opts.json,
+      });
+      process.exit(result.exitCode);
     });
 
   return command;
